#137: Move proof FetchSource structure definition into proofdb.
