+++ crev-lib/src/tests/issues.rs
@@ -1,6 +1,6 @@
 use super::*;
 
-use crate::local::FetchSource;
+use crate::proofdb::FetchSource;
 use crev_data::{
     proof,
     review::{Advisory, Issue, VersionRange},
